requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-deepseek==0.1.4
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.2
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.2
openai==1.98.0
python-dotenv==1.0.1
uuid==1.30